+++ CHANGELOG.md
@@ -6,6 +6,7 @@
  - Added `no-elif-return` linter check
  - Added `no-else-return` linter check
  - Added `gd2py` (GDScript-to-python) converter tool
+ - Added `gdradon` code metrics tool
 
 ### Fixed
  - Workarounded a lark error impacting pretty print in frontends